+++ display.h
@@ -1,7 +1,7 @@
 #ifndef H_DISPLAY
 #define H_DISPLAY
 
-void display_tab(float *tab, int N);
-void display_tab_res(float *tab, int N);
+void display_tab(double *tab, int N);
+void display_tab_res(double *tab, int N);
 
 #endif
