+++ testing.h
@@ -1,7 +1,7 @@
 #ifndef H_TESTING
 #define H_TESTING
 
-float error(float tab[], float res[], int N);
+double error(double tab[], double res[], int N);
 void test_gauss();
 void test_cholesky();
